# [XLA] vectorize row reduction for even row size