@nouiz 2020-03-04
I did the change or answer all comments.

Just a note that for odd row size, this is also complicated. There is memory order optimization problem in LLVM. I already fixed one, but the problem left is probably more complicated. I'm still trying a few different options before trying to change LLVM optimization passes.