@nouiz 2020-03-12
Can you share one HLO graph that regress with this?
Also, on which GPU do you saw regression? I didn't benchmark on many GPUs or on older GPU generations. Here is an HLO that give me speed up the fused reduction in the optimized graph:

```
HloModule cluster_0__XlaCompiledKernel_true__XlaNumConstantArgs_0__XlaNumResourceArgs_0_.25

%max_ {
  %x = f16[] parameter(0)
  %y = f16[] parameter(1)
  ROOT %maximum = f16[] maximum(%x, %y)
}

%add_ {
  %x = f32[] parameter(0)
  %y = f32[] parameter(1)
  ROOT %add = f32[] add(%x, %y)
}

ENTRY %main {
  %arg0.1 = f16[32,131070]{1,0} parameter(0)
  %constant.3 = f16[] constant(-inf)
  %pad1 = f16[32,131070] pad(%arg0.1, %constant.3), padding=0_0_0x0_0_0
  %reduce.8 = f16[32]{0} reduce(%pad1, %constant.3), dimensions={1}, to_apply=%max_
  %broadcast.9 = f16[32,131070]{1,0} broadcast(%reduce.8), dimensions={0}
  %subtract.131070 = f16[32,131070]{1,0} subtract(%arg0.1, %broadcast.9)
  %exponential.11 = f16[32,131070]{1,0} exponential(%subtract.131070)
  %convert.12 = f32[32,131070]{1,0} convert(%exponential.11)
  %constant.13 = f32[] constant(0)
  %pad2 = f32[32,131070] pad(%convert.12, %constant.13), padding=0_0_0x0_0_0
  %reduce.18 = f32[32]{0} reduce(%pad2, %constant.13), dimensions={1}, to_apply=%add_
  %convert.19 = f16[32]{0} convert(%reduce.18)
  %broadcast.20 = f16[32,131070]{1,0} broadcast(%convert.19), dimensions={0}
  ROOT %divide.21 = f16[32,131070]{1,0} divide(%exponential.11, f16[32,131070]{1,0} %broadcast.20)
}
```

Here is the nvprof output:

```
==> OUT.softmaxf16_32x131070_131070._nvprof_TFV=old_master_xla=1.txt <==
2020-03-12 13:04:20.388311: E tensorflow/compiler/xla/tools/replay_computation.cc:476] iteration complete.
==9320== Profiling application: .../tensorflow-source.copy//bazel-out/k8-opt/bin/tensorflow/compiler/xla/tools/replay_computation_gpu softmax_template.txt.tmp --print_result=false --use_fake_data --num_runs=5
==9320== Profiling result:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   90.81%  4.8267ms         1  4.8267ms  4.8267ms  4.8267ms  [CUDA memcpy DtoH]
                    3.75%  199.14us         5  39.827us  39.328us  40.607us  fusion_1
                    2.77%  147.36us         5  29.472us  29.024us  29.792us  fusion
                    2.08%  110.56us         5  22.112us  21.600us  22.432us  reduce_8
                    0.34%  17.856us         1  17.856us  17.856us  17.856us  broadcast_2
                    0.25%  13.536us        10  1.3530us  1.0880us  1.8560us  [CUDA memset]

==> OUT.softmaxf16_32x131070_131070._nvprof_TFV=rb2_xla=1.txt <==
2020-03-12 12:57:39.328761: E tensorflow/compiler/xla/tools/replay_computation.cc:476] iteration complete.
==8924== Profiling application: .../tensorflow-source.copy//bazel-out/k8-opt/bin/tensorflow/compiler/xla/tools/replay_computation_gpu softmax_template.txt.tmp --print_result=false --use_fake_data --num_runs=5
==8924== Profiling result:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   94.13%  7.3283ms         1  7.3283ms  7.3283ms  7.3283ms  [CUDA memcpy DtoH]
                    2.17%  169.22us         5  33.843us  32.608us  35.040us  fusion_1
                    1.88%  146.72us         5  29.343us  28.896us  29.791us  fusion
                    1.42%  110.24us         5  22.047us  21.760us  22.176us  reduce_8
                    0.23%  17.792us         1  17.792us  17.792us  17.792us  broadcast_2
                    0.17%  13.312us        10  1.3310us  1.0560us  1.8880us  [CUDA memset]
```