@nouiz 2020-03-16
I added 4 commits:
- The first one trigger some of the missing vectorization by LLVM. ptxas was vectorizing those already, but at least, now they are tested here.
- The 3rd one fix a regression. We where using the columns reduction indexing in some cases.
- The 2nd and 4th limit the vectorization on P100 GPUs. We now only vectorize when the tile fit exactly and for dtype of up to 32 bits. 

On P100, with float16, I see 1.23x and 2.33x speed up. In float32, I see 1.08x and 1.10x speed up.

The PR can be benchmarked. But I still need to update the tests to have expect the right vectorization depending of the hardware. I need to find how to detect the current GPU in the test.