@nouiz 2020-03-16
I updated the tests to expect the right PTX depending of the GPU arch.
I also fixed crashed that I introduced today by an earlier in d9ae69f04ba944384d117662b888c43cb7e0bf72

I did a small refactoring to help that kind of error to not pop up again here.