@cheshire 2020-03-26
@nouiz I'm a bit confused by your benchmark, do you mean the runtime reduction in `fusion_1` from `39us` to `33us`?