@nouiz 2020-03-30
Fixed. I needed to check that the layout is row major for one of the simplification.