@nouiz 2020-04-01
New PR with the CHECK fix and the other small code changes:
https://github.com/tensorflow/tensorflow/pull/38136